move tag logic to helper
