+++ app/views/people/index.html.erb
@@ -17,8 +17,8 @@
             <h3>Find people by common skills and interests</h3>
             <h4>Skills</h4>
             <p>
-              <% if Person.all_skill_tags.size > 0 %>
-                <% Person.all_skill_tags.each do |tag| %>
+              <% if skill_tags.any? %>
+                <% skill_tags.each do |tag| %>
                   <span class="tags"><%= link_to tag.name, people_with_interests_path(tag: tag.name) %></span>
                 <% end %>
               <% else %>
@@ -27,8 +27,8 @@
             </p>
             <h4>Interests</h4>
             <p>
-              <% if Person.all_interest_tags.size > 0 %>
-                <% Person.all_interest_tags.each do |tag| %>
+              <% if interest_tags.any? %>
+                <% interest_tags.each do |tag| %>
                   <span class="tags"><%= link_to tag.name, people_with_interests_path(tag: tag.name) %></span>
                 <% end %>
               <% else %>
@@ -40,8 +40,8 @@
             <h3>Search by Role</h3>
             <h4>Categories</h4>
             <p>
-              <% Person.all_person_role_tags.each do |tag| %>
-                <span class="tags"><%= link_to tag['name'], people_with_role_path(tag: tag['name']) %></span>
+              <% person_role_tags.each do |tag| %>
+                <span class="tags"><%= link_to tag.name, people_with_role_path(tag: tag.name) %></span>
               <% end %>
             </p>
           </div>
